ex01 doing

+++ CPP_Module/CPP_module_00/ex00/megaphone.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <string>
-#include <cctype>
+#include <locale>
 
 int main(int argc, char *argv[])
 {
@@ -11,7 +11,7 @@ int main(int argc, char *argv[])
         for (int i = 1; i < argc; i++)
         {
             std::string str(argv[i]);
-            for (int j = 0; j < str.length(); j++)
+            for (int j = 0; j < (int)str.length(); j++)
                 std::cout << (char)std::toupper(str[j]);
         }
         std::cout << std::endl;
